+++ src/hitchcock/urls.py
@@ -20,8 +20,6 @@ from django.views.generic import RedirectView
 
 
 urlpatterns = [
-    # Skip the uploads app index page and go directly to the changelist
-    path('admin/uploads/', RedirectView.as_view(url='upload/')),
     path('admin/', admin.site.urls),
     path('', include('uploads.urls')),
 ]